app/src/main/java/com/anggiat/jetpackmoviecatalogue/ui/tvshow/TvShowViewModel.java
package com.anggiat.jetpackmoviecatalogue.ui.tvshow;

import androidx.lifecycle.LiveData;
import androidx.lifecycle.ViewModel;
import androidx.paging.PagedList;

import com.anggiat.jetpackmoviecatalogue.data.source.MovieRepository;
import com.anggiat.jetpackmoviecatalogue.data.source.local.entity.TvShowEntity;
import com.anggiat.jetpackmoviecatalogue.vo.Resource;

public class TvShowViewModel extends ViewModel {
    private MovieRepository movieRepository;

    public TvShowViewModel(MovieRepository movieRepository) {
        this.movieRepository = movieRepository;
    }

    public LiveData<Resource<PagedList<TvShowEntity>>> getTvShows() {
        return movieRepository.getAllTvShow();
    }

}
